requirements/base.txt
fastapi~=0.95
psycopg2~=2.9
SQLAlchemy~=2.0
alembic~=1.10
pydantic~=1.10
python-jose~=3.3
requests~=2.30
uvicorn~=0.22
passlib~=1.7
python-multipart==0.0.6
bcrypt~=4.0
python-dotenv~=0.19